requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tradingeconomics==0.3.2
pandas==2.1.3
python-dotenv==1.0.0
pydantic==2.5.0